+++ src/Mew/Interfaces/IElement.ts
@@ -7,5 +7,5 @@ export interface IElement {
     content?: string,
     line?: string
     tag: string,
-    isLogic: boolean
+    isLogic?: boolean
 }