+++ src/Mew/Types/DataType.ts
@@ -1,1 +1,3 @@
-export type DataType = { [key: string]: any }
+import {Variable} from "../Logic/Variable";
+
+export type DataType = Variable[]