+++ src/Mew/Variables.ts
@@ -2,26 +2,33 @@ import {DataType} from "./Types/DataType";
 import {BlockElement} from "./Logic/BlockElement";
 import {Parser} from "./Parser";
 import {Config} from "../Mew";
+import {cl} from "../cl";
+import {Variable} from "./Logic/Variable";
 
 export module Variables {
-    export let Data: DataType = {}
-    export let DataBlock: DataType = {}
+    export let Data: DataType = []
+    export let DataBlock: DataType = []
 
     function getData() {
-        return {...Data, ...DataBlock}
+        let save: { [key: string]: Variable } = {}
+        let data: Variable[] = [...Data, ...DataBlock]
+        data.forEach((v: Variable) => save[v.name] = v)
+        data = []
+        for (const [, value] of Object.entries(save)) data.push(value)
+        return data
     }
 
     let Checkers = [
-        { // Replace
+        { // --------------------------------------------- Replace
             callback: (str: string) => {
-                for (const [key, value] of Object.entries(getData())) {
-                    const regex = new RegExp("{{([ ]+)?(" + key + ")([ ]+)?}}", "g")
-                    str = str.replace(regex, value);
-                }
+                getData().forEach((v: Variable) => {
+                    const regex = new RegExp("{{([ ]+)?(" + v.name + ")([ ]+)?}}", "g")
+                    str = str.replace(regex, v.value);
+                })
                 return str;
             }
         },
-        { // Set or Change value
+        { // --------------------------------------------- Set or Change value
             callback: (str: string) => {
                 const regex = /\$([\w]+)[ ]*?([=])(.*)/
                 str = str.trim()
@@ -31,13 +38,13 @@ export module Variables {
                     let vName: string = m[1]
                     let vValue: string = m[3].trim()
                     if (vValue.charAt(0) === '"') vValue = vValue.substr(1, vValue.length - 2)
-                    Data[vName] = vValue;
+                    Variables.addVariable(vName, vValue)
                     str = ""
                 }
                 return str;
             }
         },
-        { // For in
+        { // --------------------------------------------- For in
             callback: (str: string) => {
                 const regex = /for ([\w]+) in ([\w]+)/
                 str = str.trim()
@@ -47,16 +54,36 @@ export module Variables {
                     if (m.index === regex.lastIndex) regex.lastIndex++;
                     let vIteration: string | any = m[2]
                     let vValue: string = m[1]
-                    for (const [key, value] of Object.entries(getData())) {
-                        if (key === vIteration) {
+                    getData().forEach((v: Variable) => {
+                        if (v.name === vIteration) {
                             str = `loop-for(iteration="${vIteration}" value="${vValue}")`
                         }
-                    }
+                    })
                 }
                 return str;
             }
         }
     ]
+
+    export const addVariable = (name: string, value: any, type: string | undefined = undefined) => {
+        Data.push(new Variable(name, value, type))
+    }
+
+    export const addVariables = (variables: Object) => {
+        for (let [key, value] of Object.entries(variables)) addVariable(key, value)
+    }
+
+    export const addBlockVariable = (name: string, value: any, type: string | undefined = undefined) => {
+        DataBlock.push(new Variable(name, value, type))
+    }
+
+    export const addBlockVariables = (variables: Object) => {
+        for (let [key, value] of Object.entries(variables)) addBlockVariable(key, value)
+    }
+
+    export const blockReset = () => {
+        DataBlock = []
+    }
 
     export const check = (str: string) => {
         Checkers.forEach(checker => str = checker.callback(str))