+++ tests/main.js
@@ -4,12 +4,18 @@ Mew.Config = {
     entry_file: "./tests/index",
     variables: {
         bonjour: "Hello world from Mew! ♥",
+        amount: "5",
         fruits: [
             "bananes",
             "pommes",
             "poires",
             "fraises"
-        ]
+        ],
+        user: {
+            name: 'jean',
+            age: 28,
+            is_online: true
+        }
     }
 }